+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.6.0] - 2019-07-19
+
+- Added
+  - Auto-refresh tree view when translating objects
+  - Generating signed URLs for uploaded objects
+  - Viewing object manifests
+- Fixed
+  - Error when viewing properties of objects with no properties.
+
 ## [0.5.1] - 2019-07-18
 
 - Changed